refactor(tests): test_terraform_streaming.py を tests/streaming/ 配下の責務別ファイルに分割 (#426)
issue #426 の方針に従い、3780 行 / 26 クラスの巨大テストファイルを責務単位で
分割。テスト本体のロジックは無変更で移動のみ。

分割後の構成 (tests/streaming/):

- _helpers.py        — 共有パス定数 + _extract_yaml_packages_block
- test_terraform_meta.py    — versions.tf / outputs.tf / tfvars / .gitignore / install_root テンプレ
- test_variables_tf.py      — variables.tf 群 (基本 + null_resource + firewall)
- test_main_tf.py           — main.tf 群 (vultr_instance / user_data / null_resource / locals / firewall / run-ffmpeg provisioner)
- test_cloud_init.py        — cloud-init.yaml + env.tftpl
- test_systemd_unit.py      — youtube-stream.service.tftpl
- test_scripts.py           — swap_video.sh + run-ffmpeg.sh
- test_skill_streaming.py   — streaming SKILL.md / README.md

テスト件数は 215 件で分割前後一致。
